# models/apiDocumentation/openapi.yaml
openapi: 3.0.0
info:
  title: SFL
  contact: {}
  version: '1.0'
servers:
- url: https://sfl-clause-analysis-01.herokuapp.com
  variables: {}
paths:
...
  /group/createGroup/Arti001:
    post:
      tags:
      - Groups
      summary: Create Group
      operationId: CreateGroup
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
            example:
              groupName: Maynooth English Literature
              groupID: 4
              groupCapacity: 5
        required: true
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /group/joinGroup/100/arti@071:
    put:
      tags:
      - Groups
      summary: Join Group
      operationId: JoinGroup
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /group/allGroups:
    get:
      tags:
      - Groups
      summary: All Groups
      operationId: AllGroups
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /group/deleteGroup/112222:
    delete:
      tags:
      - Groups
      summary: Delete Group
      operationId: DeleteGroup
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /group/getAllMembers/123456:
    get:
      tags:
      - Groups
      summary: Get All Members
      operationId: GetAllMembers
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
...
  /group/getUserID:
    get:
      tags:
      - Groups
      summary: get UserID
      operationId: getUserID
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /group/getGroupID/group_1:
    get:
      tags:
      - Groups
      summary: Get groupID
      operationId: GetgroupID
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
...
  /analysis/allAnalysis:
    get:
      tags:
      - Analysis
      summary: All Analysis
      operationId: AllAnalysis
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
...
  /group/Updateinvite/1/buuzuu19:
    put:
      tags:
      - Invite
      summary: Update Invite
      operationId: UpdateInvite
      parameters:
      - name: x-auth-token
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhyaXRpa2d1cHRhMjRAZ21haWwuY29tIiwicGFzc3dvcmQiOiIxMjMxMjMiLCJpYXQiOjE2MjIyNDMwMzd9.TiDIhTUWMtGiss9hDGqLsgMG0zW642j_SKapjWmgG4M
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
...
components:
  schemas:
...
    CreateGroupRequest:
      title: CreateGroupRequest
      required:
      - groupName
      - groupID
      - groupCapacity
      type: object
      properties:
        groupName:
          type: string
        groupID:
          type: integer
          format: int32
        groupCapacity:
          type: integer
          format: int32
      example:
        groupName: Maynooth English Literature
        groupID: 4
        groupCapacity: 5
...
  securitySchemes:
    apiKey:
      type: apiKey
      name: x-auth-token
      in: header
security: []
...